+++ app/Http/Controllers/IndexView.php
@@ -9,13 +9,11 @@ use Illuminate\Support\Str;
 use SplFileObject;
 
 /**
- * Manage items
- *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class IndexController extends Controller
+class IndexView extends Controller
 {
     /**
      * Return all routes
